Fix mockHistoryArchiveAdapter is unused error
