+++ services/horizon/internal/ingest/history_archive_adapter_test.go
@@ -7,8 +7,30 @@ import (
 	"testing"
 
 	"github.com/stellar/go/historyarchive"
+	"github.com/stellar/go/ingest/io"
+	"github.com/stellar/go/xdr"
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/mock"
 )
+
+type mockHistoryArchiveAdapter struct {
+	mock.Mock
+}
+
+func (m *mockHistoryArchiveAdapter) GetLatestLedgerSequence() (uint32, error) {
+	args := m.Called()
+	return args.Get(0).(uint32), args.Error(1)
+}
+
+func (m *mockHistoryArchiveAdapter) BucketListHash(sequence uint32) (xdr.Hash, error) {
+	args := m.Called(sequence)
+	return args.Get(0).(xdr.Hash), args.Error(1)
+}
+
+func (m *mockHistoryArchiveAdapter) GetState(ctx context.Context, sequence uint32) (io.ChangeReader, error) {
+	args := m.Called(ctx, sequence)
+	return args.Get(0).(io.ChangeReader), args.Error(1)
+}
 
 func TestGetState_Read(t *testing.T) {
 	archive, e := getTestArchive()
